quik: buildable under squeeze
A few large changes:
1) removed the need for ext2fs lib, the
   new ext2fs implementation comes from u-boot
   and gives symlink support for free.
2) cleaned up sources a bit, cleaning up
   the entire disk/file/part access code quite
   a bit

+++ second/prom.c
@@ -154,9 +154,9 @@ prom_shim(struct prom_args *args)
          }
 
          if (strcmp(name, "linux,initrd-start") == 0) {
-            *place = of_shim_state.initrd_base;
+           *place = (uint32_t) of_shim_state.initrd_base;
          } else if (strcmp(name, "linux,initrd-end") == 0) {
-            *place = of_shim_state.initrd_base +
+           *place = (uint32_t) of_shim_state.initrd_base +
                of_shim_state.initrd_len;
          } else {
             goto out;
